Fixed issue in __setattr__ / __setitem__ of the set classes.
